+++ AppCenter/AppCenter/Internals/Util/MSConstants+Internal.h
@@ -18,7 +18,6 @@ static NSString *const kMSHeaderContentTypeKey = @"Content-Type";
 static NSString *const kMSAppCenterContentType = @"application/json";
 static NSString *const kMSHeaderContentEncodingKey = @"Content-Encoding";
 static NSString *const kMSHeaderContentEncoding = @"gzip";
-static NSString *const kMSAuthorizationHeaderKey = @"Authorization";
 static NSString *const kMSRetryHeaderKey = @"x-ms-retry-after-ms";
 
 // Token obfuscation.
@@ -34,19 +33,8 @@ static NSString *const kMSCFBundleURLTypes = @"CFBundleURLTypes";
 static NSString *const kMSCFBundleURLSchemes = @"CFBundleURLSchemes";
 static NSString *const kMSCFBundleTypeRole = @"CFBundleTypeRole";
 
-/**
- * The key for auth token history array stored in keychain.
- */
-static NSString *const kMSAuthTokenHistoryKey = @"MSAppCenterAuthTokenHistory";
-
-/**
- * Maximum amount of available token stored in the keychain.
- */
-static int const kMSMaxAuthTokenArraySize = 5;
-
 // Other HTTP constants.
 static short const kMSHTTPMinGZipLength = 1400;
-static NSString *const kMSBearerTokenHeaderFormat = @"Bearer %@";
 
 /**
  * Enum indicating result of a MSIngestionCall.