+++ AppCenterDistribute/AppCenterDistribute/Internals/Ingestion/MSDistributeIngestion.m
@@ -14,27 +14,13 @@
 /**
  * The API paths for latest release requests.
  */
-static NSString *const kMSLatestPrivateReleaseApiPathFormat = @"/sdk/apps/%@/releases/latest";
-static NSString *const kMSLatestPublicReleaseApiPathFormat = @"/public/sdk/apps/%@/distribution_groups/%@/releases/latest";
+static NSString *const kMSLatestPrivateReleaseApiPathFormat = @"/sdk/apps/%@/releases/private/latest";
+static NSString *const kMSLatestPublicReleaseApiPathFormat = @"/public/sdk/apps/%@/releases/latest";
 
-- (id)initWithHttpClient:(id<MSHttpClientProtocol>)httpClient
-                 baseUrl:(NSString *)baseUrl
-               appSecret:(NSString *)appSecret
-             updateToken:(NSString *)updateToken
-     distributionGroupId:(NSString *)distributionGroupId
-            queryStrings:(NSDictionary *)queryStrings {
-  NSString *apiPath;
-  NSDictionary *header = nil;
-  if (updateToken) {
-    apiPath = [NSString stringWithFormat:kMSLatestPrivateReleaseApiPathFormat, appSecret];
-    header = @{kMSHeaderUpdateApiToken : updateToken};
-  } else {
-    apiPath = [NSString stringWithFormat:kMSLatestPublicReleaseApiPathFormat, appSecret, distributionGroupId];
-  }
-  if ((self = [super initWithHttpClient:httpClient baseUrl:baseUrl apiPath:apiPath headers:header queryStrings:queryStrings])) {
+- (id)initWithHttpClient:(id<MSHttpClientProtocol>)httpClient baseUrl:(NSString *)baseUrl appSecret:(NSString *)appSecret {
+  if ((self = [super initWithHttpClient:httpClient baseUrl:baseUrl apiPath:nil headers:nil queryStrings:nil])) {
     _appSecret = appSecret;
   }
-
   return self;
 }
 
@@ -42,7 +28,7 @@ static NSString *const kMSLatestPublicReleaseApiPathFormat = @"/public/sdk/apps/
   return kMSHttpMethodGet;
 };
 
-- (NSDictionary *)getHeadersWithData:(NSObject *__unused)data eTag:(NSString *)eTag authToken:(NSString *__unused)authToken {
+- (NSDictionary *)getHeadersWithData:(NSObject *__unused)data eTag:(NSString *)eTag {
 
   // Set Header params.
   NSMutableDictionary *headers = [self.httpHeaders mutableCopy];
@@ -82,6 +68,25 @@ static NSString *const kMSLatestPublicReleaseApiPathFormat = @"/public/sdk/apps/
     MSLogVerbose([MSAppCenter logTag], @"URL: %@", url);
     MSLogVerbose([MSAppCenter logTag], @"Headers: %@", [flattenedHeaders componentsJoinedByString:@", "]);
   }
+}
+
+#pragma mark - MSDistributeIngestion
+
+- (void)checkForPublicUpdateWithQueryStrings:(NSDictionary *)queryStrings
+                           completionHandler:(MSSendAsyncCompletionHandler)completionHandler {
+  self.httpHeaders = @{};
+  self.apiPath = [NSString stringWithFormat:kMSLatestPublicReleaseApiPathFormat, self.appSecret];
+  self.sendURL = [super buildURLWithBaseURL:self.baseURL apiPath:self.apiPath queryStrings:queryStrings];
+  [self sendAsync:nil completionHandler:completionHandler];
+}
+
+- (void)checkForPrivateUpdateWithUpdateToken:(NSString *)updateToken
+                                queryStrings:(NSDictionary *)queryStrings
+                           completionHandler:(MSSendAsyncCompletionHandler)completionHandler {
+  self.httpHeaders = @{kMSHeaderUpdateApiToken : updateToken};
+  self.apiPath = [NSString stringWithFormat:kMSLatestPrivateReleaseApiPathFormat, self.appSecret];
+  self.sendURL = [super buildURLWithBaseURL:self.baseURL apiPath:self.apiPath queryStrings:queryStrings];
+  [self sendAsync:nil completionHandler:completionHandler];
 }
 
 @end