+++ Sasquatch/Sasquatch/Sasquatch-Bridging-Header.h
@@ -11,15 +11,9 @@
 #import "CrashLib.h"
 
 #if GCC_PREPROCESSOR_MACRO_PUPPET
-#import "MSDictionaryDocument.h"
 #import "MSEventFilter.h"
 #import "MSEventPropertiesInternal.h"
-#import "MSPaginatedDocuments.h"
-#import "MSUserInformation.h"
-#import "MSWriteOptions.h"
 #else
 @import AppCenter;
 @import AppCenterAnalytics;
-@import AppCenterData;
-@import AppCenterAuth;
 #endif